drivers/periph/timer: Use uint32_t for frequency
For all currently supported platforms `unsigned long` is 32 bit in width. But
better use `uint32_t` to be safe.

+++ cpu/fe310/periph/timer.c
@@ -37,7 +37,7 @@ static timer_cb_t isr_cb;
  */
 static void *isr_arg;
 
-int timer_init(tim_t dev, unsigned long freq, timer_cb_t cb, void *arg)
+int timer_init(tim_t dev, uint32_t freq, timer_cb_t cb, void *arg)
 {
     /* Using RISC-V built in timer (64bit value) */
     if (dev != 0) {
